ПикЯП/ConnectionTGBot/requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==21.8
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
